Extract and fix common issue status menu component
