+++ src/components/DetailView/DetailView.tsx
@@ -2,20 +2,14 @@ import {
   Accordion,
   Box,
   Breadcrumbs,
-  Button,
   Group,
-  Menu,
   Paper,
   ScrollArea,
   Stack,
   Text,
   Title,
 } from "@mantine/core"
-import { IconCaretDown } from "@tabler/icons"
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query"
 import { Issue } from "types"
-import { useState } from "react"
-import { getIssueTypes, setStatus } from "../CreateIssue/queryFunctions"
 import { AddSubtask } from "./Components/AddSubtask"
 import { AssigneeMenu } from "./Components/AssigneeMenu"
 import { EditableEpic } from "./Components/EditableEpic"
@@ -31,8 +25,7 @@ import { DeleteIssue } from "./Components/DeleteIssue"
 import { Attachments } from "./Components/Attachments/Attachments"
 import { ColorSchemeToggle } from "../common/ColorSchemeToggle"
 import { IssueIcon } from "../BacklogView/Issue/IssueIcon"
-
-import classes from './DetailView.module.css';
+import { IssueStatusMenu } from "./Components/IssueStatusMenu";
 
 export function DetailView({
   issueKey,
@@ -53,22 +46,6 @@ export function DetailView({
   attachments,
   closeModal,
 }: Issue & { closeModal: () => void }) {
-  const [opened, setOpened] = useState(false)
-
-  const { data: issueTypes } = useQuery({
-    queryKey: ["issueTypes", projectId],
-    queryFn: () => getIssueTypes(projectId),
-    enabled: !!projectId,
-  })
-
-  const queryClient = useQueryClient()
-  const [defaultStatus, setDefaultStatus] = useState(status)
-  const statusMutation = useMutation({
-    mutationFn: (targetStatus: string) => setStatus(issueKey, targetStatus),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["issues"] })
-    },
-  })
 
   return (
     <Paper p="xs">
@@ -122,35 +99,8 @@ export function DetailView({
         </Stack>
         <ScrollArea.Autosize style={{ minWidth: "260px", maxHeight: "70vh", flex: 10 }}>
           <Box>
-            <Group justify="apart" mb="sm">
-              <Menu
-                shadow="md"
-                onOpen={() => setOpened(true)}
-                onClose={() => setOpened(false)}
-              >
-                <Menu.Target>
-                  <Button rightSection={<IconCaretDown className={classes.icon} />}>
-                    {defaultStatus}
-                  </Button>
-                </Menu.Target>
-
-                <Menu.Dropdown>
-                  {issueTypes &&
-                    issueTypes
-                      .find((issueType) => issueType.name === type)
-                      ?.statuses?.map((issueStatus) => (
-                        <Menu.Item
-                          key={issueStatus.id}
-                          onClick={() => {
-                            statusMutation.mutate(issueStatus.name)
-                            setDefaultStatus(issueStatus.name)
-                          }}
-                        >
-                          {issueStatus.name}
-                        </Menu.Item>
-                      ))}
-                </Menu.Dropdown>
-              </Menu>
+            <Group justify="space-between" mb="sm">
+              <IssueStatusMenu projectId={projectId} issueKey={issueKey} type={type} status={status} />
               <DeleteIssue issueKey={issueKey} closeModal={closeModal} />
             </Group>
             <Accordion variant="contained" defaultValue="Details" mb={20}>
